[ReflectionProbeEditor] Influence Normal edition
